Move `FioRequestRow` to the right folder
